allow disabling of device reboots globally
Certain device types may run Android within a chroot environment,
which does not support device reboot very well. While tests typically
don't require reboot as an integral step in testing, exceptions
include reboot stress test which won't be applicable for devices in
this class.

This change enables tests to be executed on a local basis without the
need to update each test.

+++ src/com/android/tradefed/device/AndroidNativeDevice.java
@@ -2449,6 +2449,10 @@ public class AndroidNativeDevice implements IManagedTestDevice {
             CLog.i("device %s in fastboot. Rebooting to userspace.", getSerialNumber());
             executeFastbootCommand("reboot");
         } else {
+            if (mOptions.shouldDisableReboot()) {
+                CLog.i("Device reboot disabled by options, skipped.");
+                return;
+            }
             CLog.i("Rebooting device %s", getSerialNumber());
             doAdbReboot(null);
             waitForDeviceNotAvailable("reboot", DEFAULT_UNAVAILABLE_TIMEOUT);
